requirements.txt
apex==0.9.10.dev0
Augmentor==0.2.9
cloudpickle==1.4.1
hyperopt>=0.2.7
matplotlib>=3.5.1
numba>=0.55.0
numpy>=1.21.4,<1.22
pandas>=1.3.5
Pillow>=8.4.0
pretrainedmodels==0.7.4
pytorch-warmup==0.0.4
scikit-learn>=1.0.1
scipy>=1.7.3
skorch==0.9.0
timm==0.5.4
torch==1.6.0
torch_lr_finder==0.2.1
torchvision==0.7.0
tqdm>=4.62.3
https://github.com/juntang-zhuang/Adabelief-Optimizer/archive/update_0.3.0.zip#subdirectory=pypi_packages/adabelief_pytorch0.3.0
https://github.com/llllllllll/cloudpickle-generators/archive/master.zip
https://github.com/cebtenzzre/typed-argument-parser/archive/master.zip
